+++ lib/clearance.rb
@@ -1,13 +1,13 @@
-require 'clearance/configuration'
-require 'clearance/sign_in_guard'
-require 'clearance/session'
-require 'clearance/rack_session'
-require 'clearance/back_door'
-require 'clearance/controller'
-require 'clearance/user'
-require 'clearance/password_strategies'
-require 'clearance/constraints'
-require 'clearance/engine'
+require "clearance/configuration"
+require "clearance/sign_in_guard"
+require "clearance/session"
+require "clearance/rack_session"
+require "clearance/back_door"
+require "clearance/controller"
+require "clearance/user"
+require "clearance/password_strategies"
+require "clearance/constraints"
+require "clearance/engine"
 
 module Clearance
 end