+++ lib/clearance/user.rb
@@ -1,7 +1,7 @@
-require 'digest/sha1'
-require 'active_model'
-require 'email_validator'
-require 'clearance/token'
+require "digest/sha1"
+require "active_model"
+require "email_validator"
+require "clearance/token"
 
 module Clearance
   # Required to be included in your configued user class, which is `User` by
@@ -113,7 +113,7 @@ module Clearance
     # @api private
     module ClassMethods
       def authenticate(email, password)
-        if user = find_by_normalized_email(email)
+        if (user = find_by_normalized_email(email))
           if password.present? && user.authenticated?(password)
             user
           end
@@ -150,9 +150,9 @@ module Clearance
 
       included do
         validates :email,
-          email: { mode: :strict },
+          email: {mode: :strict},
           presence: true,
-          uniqueness: { allow_blank: true, case_sensitive: true },
+          uniqueness: {allow_blank: true, case_sensitive: true},
           unless: :email_optional?
 
         validates :password, presence: true, unless: :skip_password_validation?