+++ lib/generators/clearance/routes/routes_generator.rb
@@ -1,9 +1,9 @@
-require 'rails/generators/base'
+require "rails/generators/base"
 
 module Clearance
   module Generators
     class RoutesGenerator < Rails::Generators::Base
-      source_root File.expand_path('../templates', __FILE__)
+      source_root File.expand_path("../templates", __FILE__)
 
       def inject_clearance_routes_into_application_routes
         route(clearance_routes)
@@ -13,7 +13,7 @@ module Clearance
         inject_into_file(
           "config/initializers/clearance.rb",
           "  config.routes = false\n",
-          after: "Clearance.configure do |config|\n",
+          after: "Clearance.configure do |config|\n"
         )
       end
 
@@ -24,7 +24,7 @@ module Clearance
       end
 
       def routes_file_path
-        File.expand_path(find_in_source_paths('routes.rb'))
+        File.expand_path(find_in_source_paths("routes.rb"))
       end
 
       def route(routing_code)
@@ -37,7 +37,7 @@ module Clearance
             routing_code,
             after: sentinel,
             verbose: false,
-            force: true,
+            force: true
           )
         end
       end