+++ spec/clearance/session_spec.rb
@@ -1,13 +1,13 @@
-require 'spec_helper'
+require "spec_helper"
 
 describe Clearance::Session do
-  before { Timecop.freeze }
-  after { Timecop.return }
+  before { freeze_time }
+  after { unfreeze_time }
 
   let(:session) { Clearance::Session.new(env_without_remember_token) }
   let(:user) { create(:user) }
 
-  it 'finds a user from a cookie' do
+  it "finds a user from a cookie" do
     user = create(:user)
     env = env_with_remember_token(user.remember_token)
     session = Clearance::Session.new(env)
@@ -16,15 +16,15 @@ describe Clearance::Session do
     expect(session.current_user).to eq user
   end
 
-  it 'returns nil for an unknown user' do
-    env = env_with_remember_token('bogus')
+  it "returns nil for an unknown user" do
+    env = env_with_remember_token("bogus")
     session = Clearance::Session.new(env)
 
     expect(session).to be_signed_out
     expect(session.current_user).to be_nil
   end
 
-  it 'returns nil without a remember token' do
+  it "returns nil without a remember token" do
     expect(session).to be_signed_out
     expect(session.current_user).to be_nil
   end
@@ -71,7 +71,7 @@ describe Clearance::Session do
 
     context "signed cookie exists" do
       it "uses cookies.signed[remember_token]" do
-        cookie_jar = { "remember_token" => "signed cookie" }
+        cookie_jar = {"remember_token" => "signed cookie"}
         expect(session).to receive(:cookies).and_return(cookie_jar)
         expect(cookie_jar).to receive(:signed).and_return(cookie_jar)
 
@@ -81,7 +81,7 @@ describe Clearance::Session do
 
     context "signed cookie does not exist yet" do
       it "uses cookies[remember_token] instead" do
-        cookie_jar = { "remember_token" => "signed cookie" }
+        cookie_jar = {"remember_token" => "signed cookie"}
         # first call will try to get the signed cookie
         expect(session).to receive(:cookies).and_return(cookie_jar)
         # ... but signed_cookie doesn't exist
@@ -94,8 +94,8 @@ describe Clearance::Session do
     end
   end
 
-  describe '#sign_in' do
-    it 'sets current_user' do
+  describe "#sign_in" do
+    it "sets current_user" do
       user = build(:user)
 
       session.sign_in user
@@ -103,8 +103,8 @@ describe Clearance::Session do
       expect(session.current_user).to eq user
     end
 
-    context 'with a block' do
-      it 'passes the success status to the block when sign in succeeds' do
+    context "with a block" do
+      it "passes the success status to the block when sign in succeeds" do
         success_status = stub_status(Clearance::SuccessStatus, true)
         success_lambda = stub_callable
 
@@ -113,7 +113,7 @@ describe Clearance::Session do
         expect(success_lambda).to have_been_called.with(success_status)
       end
 
-      it 'passes the failure status to the block when sign in fails' do
+      it "passes the failure status to the block when sign in fails" do
         failure_status = stub_status(Clearance::FailureStatus, false)
         failure_lambda = stub_callable
 
@@ -135,17 +135,16 @@ describe Clearance::Session do
       end
     end
 
-    context 'with nil argument' do
-      it 'assigns current_user' do
+    context "with nil argument" do
+      it "assigns current_user" do
         session.sign_in nil
 
         expect(session.current_user).to be_nil
       end
     end
 
-    context 'with a sign in stack' do
-
-      it 'runs the first guard' do
+    context "with a sign in stack" do
+      it "runs the first guard" do
         guard = stub_sign_in_guard(succeed: true)
         user = build(:user)
 
@@ -154,13 +153,13 @@ describe Clearance::Session do
         expect(guard).to have_received(:call)
       end
 
-      it 'will not sign in the user if the guard stack fails' do
+      it "will not sign in the user if the guard stack fails" do
         stub_sign_in_guard(succeed: false)
         user = build(:user)
 
         session.sign_in user
 
-        expect(session.instance_variable_get("@cookies")).to be_nil
+        expect(session.instance_variable_get(:@cookies)).to be_nil
         expect(session.current_user).to be_nil
       end
 
@@ -174,23 +173,23 @@ describe Clearance::Session do
 
       def stub_default_sign_in_guard
         double("default_sign_in_guard").tap do |sign_in_guard|
-          allow(Clearance::DefaultSignInGuard).to receive(:new).
-            with(session).
-            and_return(sign_in_guard)
+          allow(Clearance::DefaultSignInGuard).to receive(:new)
+            .with(session)
+            .and_return(sign_in_guard)
         end
       end
 
       def stub_guard_class(guard)
         double("guard_class").tap do |guard_class|
-          allow(guard_class).to receive(:to_s).
-            and_return(guard_class)
+          allow(guard_class).to receive(:to_s)
+            .and_return(guard_class)
 
-          allow(guard_class).to receive(:constantize).
-            and_return(guard_class)
+          allow(guard_class).to receive(:constantize)
+            .and_return(guard_class)
 
-          allow(guard_class).to receive(:new).
-            with(session, stub_default_sign_in_guard).
-            and_return(guard)
+          allow(guard_class).to receive(:new)
+            .with(session, stub_default_sign_in_guard)
+            .and_return(guard)
         end
       end
 
@@ -204,25 +203,25 @@ describe Clearance::Session do
     end
   end
 
-  context 'if httponly is set' do
+  context "if httponly is set" do
     before do
       session.sign_in(user)
     end
 
-    it 'sets a httponly cookie' do
+    it "sets a httponly cookie" do
       session.add_cookie_to_headers
 
       expect(remember_token_cookie(session)[:httponly]).to be_truthy
     end
   end
 
-  context 'if httponly is not set' do
+  context "if httponly is not set" do
     before do
       Clearance.configuration.httponly = false
       session.sign_in(user)
     end
 
-    it 'sets a standard cookie' do
+    it "sets a standard cookie" do
       session.add_cookie_to_headers
 
       expect(remember_token_cookie(session)[:httponly]).to be_falsey
@@ -254,9 +253,9 @@ describe Clearance::Session do
     end
   end
 
-  describe 'remember token cookie expiration' do
-    context 'default configuration' do
-      it 'is set to 1 year from now' do
+  describe "remember token cookie expiration" do
+    context "default configuration" do
+      it "is set to 1 year from now" do
         user = double("User", remember_token: "123abc")
         session = Clearance::Session.new(env_without_remember_token)
         session.sign_in user
@@ -266,15 +265,15 @@ describe Clearance::Session do
       end
     end
 
-    context 'configured with lambda taking one argument' do
-      it 'it can use other cookies to set the value of the expires token' do
+    context "configured with lambda taking one argument" do
+      it "it can use other cookies to set the value of the expires token" do
         remembered_expires = 12.hours.from_now
         expires_at = ->(cookies) do
-          cookies['remember_me'] ? remembered_expires : nil
+          cookies["remember_me"] ? remembered_expires : nil
         end
         with_custom_expiration expires_at do
           user = double("User", remember_token: "123abc")
-          environment = env_with_cookies(remember_me: 'true')
+          environment = env_with_cookies(remember_me: "true")
           session = Clearance::Session.new(environment)
           session.sign_in user
           session.add_cookie_to_headers
@@ -287,26 +286,26 @@ describe Clearance::Session do
     end
   end
 
-  describe 'secure cookie option' do
-    context 'when not set' do
+  describe "secure cookie option" do
+    context "when not set" do
       before do
         session.sign_in(user)
       end
 
-      it 'sets a standard cookie' do
+      it "sets a standard cookie" do
         session.add_cookie_to_headers
 
         expect(remember_token_cookie(session)[:secure]).to be_falsey
       end
     end
 
-    context 'when set' do
+    context "when set" do
       before do
         Clearance.configuration.secure_cookie = true
         session.sign_in(user)
       end
 
-      it 'sets a secure cookie' do
+      it "sets a secure cookie" do
         session.add_cookie_to_headers
 
         expect(remember_token_cookie(session)[:secure]).to be_truthy
@@ -342,10 +341,10 @@ describe Clearance::Session do
       end
     end
 
-    context 'when not set' do
+    context "when not set" do
       before { session.sign_in(user) }
 
-      it 'sets a standard cookie' do
+      it "sets a standard cookie" do
         session.add_cookie_to_headers
 
         expect(remember_token_cookie(session)[:domain]).to be_nil
@@ -353,24 +352,24 @@ describe Clearance::Session do
     end
   end
 
-  describe 'cookie path option' do
-    context 'when not set' do
+  describe "cookie path option" do
+    context "when not set" do
       before { session.sign_in(user) }
 
-      it 'sets a standard cookie' do
+      it "sets a standard cookie" do
         session.add_cookie_to_headers
 
         expect(remember_token_cookie(session)[:domain]).to be_nil
       end
     end
 
-    context 'when set' do
+    context "when set" do
       before do
-        Clearance.configuration.cookie_path = '/user'
+        Clearance.configuration.cookie_path = "/user"
         session.sign_in(user)
       end
 
-      it 'sets a standard cookie' do
+      it "sets a standard cookie" do
         session.add_cookie_to_headers
 
         expect(remember_token_cookie(session)[:path]).to eq("/user")
@@ -378,7 +377,7 @@ describe Clearance::Session do
     end
   end
 
-  it 'does not set a remember token when signed out' do
+  it "does not set a remember token when signed out" do
     session = Clearance::Session.new(env_without_remember_token)
     session.add_cookie_to_headers
     expect(remember_token_cookie(session)).to be_nil
@@ -411,7 +410,7 @@ describe Clearance::Session do
         user = create(:user)
         env = env_with_remember_token(
           value: user.remember_token,
-          domain: domain,
+          domain: domain
         )
         session = Clearance::Session.new(env)
         cookie_jar = ActionDispatch::Request.new(env).cookie_jar
@@ -423,9 +422,9 @@ describe Clearance::Session do
       end
     end
 
-    context 'with callable cookie domain' do
-      it 'clears cookie' do
-        domain = '.example.com'
+    context "with callable cookie domain" do
+      it "clears cookie" do
+        domain = ".example.com"
         Clearance.configuration.cookie_domain = ->(_) { domain }
         user = create(:user)
         env = env_with_remember_token(
@@ -454,11 +453,11 @@ describe Clearance::Session do
   end
 
   def env_with_cookies(cookies)
-    Rack::MockRequest.env_for '/', 'HTTP_COOKIE' => serialize_cookies(cookies)
+    Rack::MockRequest.env_for "/", "HTTP_COOKIE" => serialize_cookies(cookies)
   end
 
   def env_with_remember_token(token)
-    env_with_cookies 'remember_token' => token
+    env_with_cookies "remember_token" => token
   end
 
   def env_without_remember_token
@@ -472,8 +471,7 @@ describe Clearance::Session do
       Rack::Utils.set_cookie_header! header, key, value
     end
 
-    cookie = header["set-cookie"] || header["Set-Cookie"]
-    cookie
+    header["set-cookie"] || header["Set-Cookie"]
   end
 
   def have_been_called