+++ spec/dummy/app/controllers/application_controller.rb
@@ -2,6 +2,6 @@ class ApplicationController < ActionController::Base
   include Clearance::Controller
 
   def show
-    render inline: "Hello user #<%= current_user.id %>", layout: false
+    render inline: "Hello user #<%= current_user.id %> #{params.to_json}", layout: false
   end
 end