+++ spec/dummy/db/migrate/20110111224543_create_clearance_users.rb
@@ -1,4 +1,4 @@
-class CreateClearanceUsers < ActiveRecord::Migration
+class CreateClearanceUsers < ActiveRecord::Migration[Rails::VERSION::STRING.to_f]
   def self.up
     create_table :users do |t|
       t.timestamps null: false
@@ -9,7 +9,8 @@ class CreateClearanceUsers < ActiveRecord::Migration
     end
 
     add_index :users, :email
-    add_index :users, :remember_token
+    add_index :users, :confirmation_token, unique: true
+    add_index :users, :remember_token, unique: true
   end
 
   def self.down