+++ CHANGELOG.md
@@ -5,7 +5,34 @@ complete changelog, see the git history for each version via the version links.
 
 ## [Unreleased]
 
-[Unreleased]: https://github.com/thoughtbot/clearance/compare/v2.8.0...main
+[Unreleased]: https://github.com/thoughtbot/clearance/compare/v2.10.0...main
+
+## [2.10.0] - March 28, 2025
+- Drop support for Rails 7.0 and Ruby 3.1 (#1045)
+
+[2.10.0]: https://github.com/thoughtbot/clearance/compare/v2.9.3...v2.10.0
+
+## [2.9.3] - November 26, 2024
+- Add Rails 8 to testing matrix (#1038) Hamed Asghari
+
+[2.9.3]: https://github.com/thoughtbot/clearance/compare/v2.9.2...v2.9.3
+
+## [2.9.2] - November 14, 2024
+- Fix query params being clobbered by Clearance::BackDoor (#1041) Frederick Cheung
+
+[2.9.2]: https://github.com/thoughtbot/clearance/compare/v2.9.1...v2.9.2
+
+## [2.9.1] - October 29, 2024
+ - Update gemspec
+
+[2.9.1]: https://github.com/thoughtbot/clearance/compare/v2.9.0...v2.9.1
+
+## [2.9.0] - October 29, 2024
+- Added Rails 7.2 and Ruby 3.3 to testing matrix
+  and overhauled dummy app (#1032) Hamed Asghari
+- Droped support for Ruby 3.0 and Rails 6.1 (#1036)
+
+[2.9.0]: https://github.com/thoughtbot/clearance/compare/v2.8.0...v2.9.0
 
 ## [2.8.0] - August 9, 2024
 - Feature: Added allow_password_resets config option (#1019) Jos O'shea